+++ program/mkdata/modules/variables.f90
@@ -1,13 +1,13 @@
 module variables
   implicit none
-  real(4), allocatable :: bs_pre(:)
-  real(4), allocatable :: bs_tem(:)
-  real(4), allocatable :: bs_qv(:)
-  real(4), allocatable :: u_profile(:)
-  real(4), allocatable :: pre(:,:,:)
-  real(4), allocatable :: tem(:,:,:)
-  real(4), allocatable :: rho(:,:,:)
-  real(4), allocatable :: qv(:,:,:)
-  real(4), allocatable :: u(:,:,:)
-  real(4), allocatable :: v(:,:,:)
+  real(8), allocatable :: bs_pre(:)
+  real(8), allocatable :: bs_tem(:)
+  real(8), allocatable :: bs_qv(:)
+  real(8), allocatable :: u_profile(:)
+  real(8), allocatable :: pre(:,:,:)
+  real(8), allocatable :: tem(:,:,:)
+  real(8), allocatable :: rho(:,:,:)
+  real(8), allocatable :: qv(:,:,:)
+  real(8), allocatable :: u(:,:,:)
+  real(8), allocatable :: v(:,:,:)
 end module variables
